Move Kibana namer from kibana/name/name.go to kibana/name.go
