+++ pkg/controller/common/keystore/user_secret_test.go
@@ -21,7 +21,7 @@ import (
 	"github.com/elastic/cloud-on-k8s/pkg/controller/common/name"
 	"github.com/elastic/cloud-on-k8s/pkg/controller/common/volume"
 	"github.com/elastic/cloud-on-k8s/pkg/controller/common/watches"
-	kbname "github.com/elastic/cloud-on-k8s/pkg/controller/kibana/name"
+	"github.com/elastic/cloud-on-k8s/pkg/controller/kibana"
 	"github.com/elastic/cloud-on-k8s/pkg/utils/k8s"
 )
 
@@ -100,7 +100,7 @@ func Test_secureSettingsVolume(t *testing.T) {
 				Watches:      tt.w,
 				FakeRecorder: record.NewFakeRecorder(1000),
 			}
-			vol, version, err := secureSettingsVolume(testDriver, &tt.kb, nil, kbname.KBNamer)
+			vol, version, err := secureSettingsVolume(testDriver, &tt.kb, nil, kibana.KBNamer)
 			require.NoError(t, err)
 			assert.Equal(t, tt.wantVolume, vol)
 			assert.Equal(t, tt.wantVersion, version)
@@ -164,7 +164,7 @@ func Test_reconcileSecureSettings(t *testing.T) {
 				userSecrets: []corev1.Secret{
 					{},
 				},
-				namer: kbname.KBNamer,
+				namer: kibana.KBNamer,
 			},
 			want:    nil,
 			wantErr: false,
@@ -180,7 +180,7 @@ func Test_reconcileSecureSettings(t *testing.T) {
 							"key1": []byte("value1"),
 						}},
 				},
-				namer: kbname.KBNamer,
+				namer: kibana.KBNamer,
 			},
 			want: &corev1.Secret{
 				ObjectMeta: expectedMeta,
@@ -206,7 +206,7 @@ func Test_reconcileSecureSettings(t *testing.T) {
 						},
 					},
 				},
-				namer: kbname.KBNamer,
+				namer: kibana.KBNamer,
 			},
 			want: &corev1.Secret{
 				ObjectMeta: expectedMeta,
@@ -226,7 +226,7 @@ func Test_reconcileSecureSettings(t *testing.T) {
 				}),
 				hasKeystore: kibanaFixture,
 				userSecrets: nil,
-				namer:       kbname.KBNamer,
+				namer:       kibana.KBNamer,
 			},
 			want:    nil,
 			wantErr: false,
@@ -237,7 +237,7 @@ func Test_reconcileSecureSettings(t *testing.T) {
 				c:           k8s.WrappedFakeClient(),
 				hasKeystore: kibanaFixture,
 				userSecrets: nil,
-				namer:       kbname.KBNamer,
+				namer:       kibana.KBNamer,
 			},
 			want:    nil,
 			wantErr: false,
@@ -259,7 +259,7 @@ func Test_reconcileSecureSettings(t *testing.T) {
 						},
 					},
 				},
-				namer: kbname.KBNamer,
+				namer: kibana.KBNamer,
 			},
 			want: &corev1.Secret{
 				ObjectMeta: expectedMeta,
@@ -287,7 +287,7 @@ func Test_reconcileSecureSettings(t *testing.T) {
 						},
 					},
 				},
-				namer: kbname.KBNamer,
+				namer: kibana.KBNamer,
 			},
 			want: &corev1.Secret{
 				ObjectMeta: expectedMeta,
