+++ test/e2e/kb/failure_test.go
@@ -7,14 +7,15 @@ package kb
 import (
 	"testing"
 
-	kbname "github.com/elastic/cloud-on-k8s/pkg/controller/kibana/name"
-	"github.com/elastic/cloud-on-k8s/test/e2e/test"
-	"github.com/elastic/cloud-on-k8s/test/e2e/test/elasticsearch"
-	"github.com/elastic/cloud-on-k8s/test/e2e/test/kibana"
 	"github.com/stretchr/testify/require"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/types"
+
+	kibana2 "github.com/elastic/cloud-on-k8s/pkg/controller/kibana"
+	"github.com/elastic/cloud-on-k8s/test/e2e/test"
+	"github.com/elastic/cloud-on-k8s/test/e2e/test/elasticsearch"
+	"github.com/elastic/cloud-on-k8s/test/e2e/test/kibana"
 )
 
 func TestKillKibanaPod(t *testing.T) {
@@ -49,7 +50,7 @@ func TestKillKibanaDeployment(t *testing.T) {
 					var dep appsv1.Deployment
 					err := k.Client.Get(types.NamespacedName{
 						Namespace: test.Ctx().ManagedNamespace(0),
-						Name:      kbname.Deployment(kbBuilder.Kibana.Name),
+						Name:      kibana2.Deployment(kbBuilder.Kibana.Name),
 					}, &dep)
 					require.NoError(t, err)
 					err = k.Client.Delete(&dep)
