+++ test/e2e/test/kibana/http_client.go
@@ -13,16 +13,17 @@ import (
 	"net/url"
 	"path"
 
-	kbv1 "github.com/elastic/cloud-on-k8s/pkg/apis/kibana/v1"
-	kbname "github.com/elastic/cloud-on-k8s/pkg/controller/kibana/name"
-	"github.com/elastic/cloud-on-k8s/test/e2e/test"
 	"github.com/pkg/errors"
+
+	kbv1 "github.com/elastic/cloud-on-k8s/pkg/apis/kibana/v1"
+	"github.com/elastic/cloud-on-k8s/pkg/controller/kibana"
+	"github.com/elastic/cloud-on-k8s/test/e2e/test"
 )
 
 func NewKibanaClient(kb kbv1.Kibana, k *test.K8sClient) (*http.Client, error) {
 	var caCerts []*x509.Certificate
 	if kb.Spec.HTTP.TLS.Enabled() {
-		crts, err := k.GetHTTPCerts(kbname.KBNamer, kb.Namespace, kb.Name)
+		crts, err := k.GetHTTPCerts(kibana.KBNamer, kb.Namespace, kb.Name)
 		if err != nil {
 			return nil, err
 		}
@@ -38,7 +39,7 @@ func DoRequest(k *test.K8sClient, kb kbv1.Kibana, password string, method string
 		scheme = "https"
 	}
 	// add .svc suffix so that requests work when using the port-forwarder during local test runs
-	u, err := url.Parse(fmt.Sprintf("%s://%s.%s.svc:5601", scheme, kbname.HTTPService(kb.Name), kb.Namespace))
+	u, err := url.Parse(fmt.Sprintf("%s://%s.%s.svc:5601", scheme, kibana.HTTPService(kb.Name), kb.Namespace))
 	if err != nil {
 		return nil, errors.Wrap(err, "while parsing url")
 	}
